finished working on products

+++ app/Http/Controllers/Product/ProductController.php
@@ -5,8 +5,9 @@ namespace App\Http\Controllers\Product;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use JWTAuth;
-use App\Http\Resources\ProductResource;
+use App\Http\Resources\Product as ProductResource;
 use App\Product;
+use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller {
     /**
@@ -16,8 +17,18 @@ class ProductController extends Controller {
     */
 
     public function index() {
-
         return ProductResource::collection( Product::with( 'user' )->paginate( 25 ) );
+
+    }
+
+    /**
+    * Show the form for creating a new resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
+
+    public function create() {
+        //
     }
 
     /**
@@ -40,7 +51,7 @@ class ProductController extends Controller {
         if ( $validator->fails() ) {
             return response()->json( $validator->errors(), 400 );
         }
-        $product_photo = $request->file( 'product_photo' );
+        $product_photo = $request->file( 'picture' );
         // File Details
         $filename = $product_photo->getClientOriginalName();
         $extension = $product_photo->getClientOriginalExtension();
@@ -62,7 +73,7 @@ class ProductController extends Controller {
                 // Upload file
                 $product_photo->move( $location, $filename );
                 $product = Product::create( [
-                    'name' => $request->name,
+                    'name' => ucfirst( $request->name ),
                     'description' => $request->description,
                     'price' => $request->price,
                     'picture' => $filename,
@@ -75,6 +86,7 @@ class ProductController extends Controller {
                 ], 201 );
             }
         }
+
         return new ProductResource( $product );
     }
 
@@ -85,8 +97,39 @@ class ProductController extends Controller {
     * @return \Illuminate\Http\Response
     */
 
-    public function show( Product $product ) {
-        return new ProductResource( $product );
+    public function show( $id ) {
+        $product = Product::find( $id );
+        if ( $product ) {
+            return new ProductResource( $product );
+        } else {
+            return response()->json( [
+                'success' => true,
+                'message' => 'Could not find a product'
+            ], 404 );
+        }
+    }
+
+    /**
+    * Show the form for editing the specified resource.
+    *
+    * @param  int  $id
+    * @return \Illuminate\Http\Response
+    */
+
+    public function edit( Request $request ) {
+        $product = Product::find( $request->product_id);
+        if ( $product ) {
+            $product->name = $request->name;
+            $product->description = $request->description;
+            $product->price = $request->price;
+            $product->save();
+            return new ProductResource( $product );
+        } else {
+            return response()->json( [
+                'success' => true,
+                'message' => 'Could not find a product'
+            ], 404 );
+        }
     }
 
     /**
@@ -97,53 +140,8 @@ class ProductController extends Controller {
     * @return \Illuminate\Http\Response
     */
 
-    public function update( Request $request, Product  $product ) {
-        $user = JWTAuth::user();
-
-        // check if currently authenticated user is the owner of the book
-        if ( $user->id !== $product->user_id ) {
-            return response()->json( ['error' => 'You can only edit your own product.'], 403 );
-        }
-
-        if ( $validator->fails() ) {
-            return response()->json( $validator->errors(), 400 );
-        }
-        $product_photo = $request->file( 'product_photo' );
-        // File Details
-        $filename = $product_photo->getClientOriginalName();
-        $extension = $product_photo->getClientOriginalExtension();
-        $tempPath = $product_photo->getRealPath();
-        $fileSize = $product_photo->getSize();
-        $mimeType = $product_photo->getMimeType();
-
-        // Valid File Extensions
-        $valid_extension = array( 'jpeg', 'png', 'gif', 'jpg' );
-        // 2MB in Bytes
-        $maxFileSize = 2097152;
-
-        // Check file extension
-        if ( in_array( strtolower( $extension ), $valid_extension ) ) {
-
-            // Check file size
-            if ( $fileSize <= $maxFileSize ) {
-                $location = 'products';
-                // Upload file
-                $product_photo->move( $location, $filename );
-                $product->update( $request->only( [
-                    'name' => $request->name,
-                    'description' => $request->description,
-                    'price' => $request->price,
-                    'picture' => $filename,
-                    ] ) );
-        
-                return new  ProductResource( $product );
-            } else {
-                return response()->json( [
-                    'error_file' => true,
-                    'message' => 'File too large. File must be less than 2MB.',
-                ], 201 );
-            }
-        }
+    public function update( Request $request, $id ) {
+        //
     }
 
     /**
@@ -153,8 +151,19 @@ class ProductController extends Controller {
     * @return \Illuminate\Http\Response
     */
 
-    public function destroy( Product $product ) {
-        $product->delete();
-        return response()->json(null, 204);
+    public function destroy( $id ) {
+        $product = Product::find( $id );
+        if ( $product ) {
+            $product->delete();
+            return response()->json( [
+                'success' => true,
+                'message' => 'Successfully deleted your shop'
+            ], 200 );
+        } else {
+            return response()->json( [
+                'success' => true,
+                'message' => 'Could not find a product'
+            ], 404 );
+        }
     }
 }
